Delete table and create table in debug screen

+++ app/src/main/java/jp/co/my/myplatform/service/content/PLDebugView.java
@@ -8,6 +8,11 @@ import android.view.View;
 import android.widget.AdapterView;
 import android.widget.TextView;
 
+import com.raizlabs.android.dbflow.config.FlowManager;
+import com.raizlabs.android.dbflow.structure.ModelAdapter;
+import com.raizlabs.android.dbflow.structure.database.DatabaseWrapper;
+
+import java.util.ArrayList;
 import java.util.List;
 
 import jp.co.my.common.util.MYLogUtil;
@@ -15,6 +20,9 @@ import jp.co.my.myplatform.R;
 import jp.co.my.myplatform.service.core.PLCoreService;
 import jp.co.my.myplatform.service.core.PLWakeLockManager;
 import jp.co.my.myplatform.service.model.PLDatabase;
+import jp.co.my.myplatform.service.model.PLNewsGroupModel;
+import jp.co.my.myplatform.service.model.PLNewsPageModel;
+import jp.co.my.myplatform.service.model.PLNewsSiteModel;
 import jp.co.my.myplatform.service.popover.PLListPopover;
 
 public class PLDebugView extends PLContentView {
@@ -33,18 +41,39 @@ public class PLDebugView extends PLContentView {
 		findViewById(R.id.delete_db_button).setOnClickListener(new OnClickListener() {
 			@Override
 			public void onClick(View v) {
-				String[] titles = {"データベース消去"};
-				new PLListPopover(titles, new AdapterView.OnItemClickListener() {
-					@Override
-					public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-						PLDebugView.this.removeTopPopover();
-						PLCoreService.getNavigationController().pushView(PLAppListView.class);
-						getContext().deleteDatabase(PLDatabase.NAME + ".db");
-						MYLogUtil.outputLog("Delete DB. Please restart app");
-					}
-				}).showPopover();
+				deleteDatabase();
 			}
 		});
+	}
+
+	private void deleteDatabase() {
+		String[] titles = {"データベース消去", "PLNewsPageModel", "PLNewsSiteModel+ PLNewsGroupModel "};
+		new PLListPopover(titles, new AdapterView.OnItemClickListener() {
+			@Override
+			public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
+				PLDebugView.this.removeTopPopover();
+				if (position == 0) {
+					PLCoreService.getNavigationController().pushView(PLAppListView.class);
+					getContext().deleteDatabase(PLDatabase.NAME + ".db");
+					MYLogUtil.outputLog("Delete DB. Please restart app");
+					return;
+				}
+
+				ArrayList<Class> klassArray = new ArrayList<>();
+				DatabaseWrapper database = FlowManager.getDatabase(PLDatabase.NAME).getHelper().getDatabase();
+				if (position == 1) {
+					klassArray.add(PLNewsPageModel.class);
+				} else {
+					klassArray.add(PLNewsSiteModel.class);
+					klassArray.add(PLNewsGroupModel.class);
+				}
+				for (Class klass : klassArray) {
+					ModelAdapter modelAdapter = FlowManager.getModelAdapter(klass);
+					database.execSQL("DROP TABLE IF EXISTS " + modelAdapter.getTableName());
+					database.execSQL(modelAdapter.getCreationQuery());
+				}
+			}
+		}).showPopover();
 	}
 
 	private void updateWakeUpText() {
